Ruta de cobro y lista de gasolina

+++ database/migrations/2020_06_11_130057_create_clients_table.php
@@ -17,8 +17,8 @@ class CreateClientsTable extends Migration
             $table->bigIncrements('id');
             $table->unsignedBigInteger('user_id');
             $table->string('membership');
-            $table->integer('current_balance');
-            $table->integer('shared_balance');
+            $table->double('current_balance');
+            $table->double('shared_balance');
             $table->integer('points');
             $table->string('image_qr');
             $table->string('birthdate')->nullable();
